fix(migration): Complete ExtraKeysPreference migration and fix method name mismatches
- Migrated ExtraKeysPreference.java to Kotlin (412→348 lines)
- Fixed method names: orderingAsAdded → setOrderingAsAdded, label_of_value → labelOfValue,
  get_serializer → getSerializer, load_from_preferences → loadFromPreferences
- Fixed LoopGestureDetector Double→Float type mismatches
- Fixed VoiceImeSwitcher nullable IBinder with safe call
- Fixed WordPredictor smart cast issue with explicit null check
- Fixed SubtypeManager method name: layout_of_string → layoutOfString
- Fixed CustomExtraKeysPreference method names to match parent class

Note: Multiple files from previous migrations still have compilation errors:
- ClipboardPinView.kt: Inheritance issue with final type
- EmojiGridView.kt: Missing OnItemClickListener import
- KeyboardReceiver.kt: Missing inputMethodManager field
- LayoutModifier.kt: Type initialization issues
- SubtypeLayoutInitializer.kt: Nullable type mismatches
- SuggestionBar.kt: Missing textColor/backgroundColor fields
- LayoutsPreference.kt: Private 'values' access and type mismatches

These require investigation of previous migration context to resolve.

+++ srcs/juloo.keyboard2/LoopGestureDetector.kt
@@ -144,8 +144,8 @@ class LoopGestureDetector(
             var angleDiff = angle2 - angle1
 
             // Normalize to [-π, π]
-            while (angleDiff > PI) angleDiff -= 2 * PI
-            while (angleDiff < -PI) angleDiff += 2 * PI
+            while (angleDiff > PI) angleDiff -= (2 * PI).toFloat()
+            while (angleDiff < -PI) angleDiff += (2 * PI).toFloat()
 
             totalAngle += angleDiff
         }
